requirements.txt
streamlit
Pillow>=9.0.0
opencv-python==4.8.0.76
ultralytics
split-folders
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
